Add remote debug capability to test project
This is a sample change to explore the idea of adding remote debug
capabilities to the kubebuilder generated scaffolding.  This is not
intended to be merged into the project as-is.  It needs to be
incorporated into the scaffolding code and tested properly.  This is for
discussion purposes only regarding PR#601.