Add wrapper script to control dlv debugger behaviour
Using the Delve debugger has a couple of shortcomings for dealing with
programs in a containerized environment:

1) The Delve debugger always waits for a client to connect before starting
the program being debugged.  This is problematic in some scenarios that
require the process be started within a certain period of time, or when
the program needs to be automatically restarted without a user being
prepared to intervene.

For this reason a wrapper script has been created that, by default,
will automatically send the debugger a "continue" signal so that the
program is started without a client being attached.

2) The Delve debugger does not exit when the program being debugged
crashes or terminates.  In a containerized environment, this means that
the container is not restarted automatically on a crash/exit.  This may
be required as a normal mode of operation in some environments.

For this reason the wrapper script runs the program in the background
rather than as a sub process of the debugger.  This way the exit of the
program can be caught and the script can exit regardless of the state of
the debugger.  This allows the container to restart as expected on
crashes.